+++ d60.Cirqus/Config/Configurers/IAggregateRootRepositoryConfigurationBuilderApi.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace d60.Cirqus.Config.Configurers
 {
-    public interface IAggregateRootRepositoryConfigurationBuilderApi : IEventDispatcherConfigurationBuilderApi
+    public interface IAggregateRootRepositoryConfigurationBuilderApi : IEventDispatcherConfigurationBuilderApi, IFullConfiguration
     {
         IEventDispatcherConfigurationBuilderApi AggregateRootRepository(Action<AggregateRootRepositoryConfigurationBuilder> configure);
     }