+++ d60.Cirqus/Config/Configurers/IEventDispatcherConfigurationBuilderApi.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace d60.Cirqus.Config.Configurers
 {
-    public interface IEventDispatcherConfigurationBuilderApi
+    public interface IEventDispatcherConfigurationBuilderApi : IFullConfiguration
     {
         IFullConfiguration EventDispatcher(Action<EventDispatcherConfigurationBuilder> configure);
     }